Google OAuth for login
Adds the ability to login to our app through a gmail account.
Making our own login system may not be the best use of time,
and will undoubtedly have security flaws. For the sake of this project,
I feel going with a good third-party solution is a good move. If we want
to make our own login system as well, I would not be opposed to learning
how to do that!