Add custom field for adding multiple authors (#2)
This was actually an interesting challenge. There is no default yesod
form field which allows you to collect a list. As a result, I had to
implement a field to do this on my own. I took the actual JS from
http://www.randomsnippets.com/2008/02/21/how-to-dynamically-add-form-elements-via-javascript/
and followed http://www.yesodweb.com/book/forms to build a custom field.