Add Reviewer and PC pages
Adds pages for submitting reviews, and for managing reviewers. Being a
reviewer adds a breadcrumb to the top where you can view all the papers
you need to review. On that page, you get a link to download the paper
and that current status, as well as any comments you left.
As the pc, you can add reviewers (deletion coming soon). This is a very
minimalistic, but working, subset of features necessary to run a
conference management system.